①在作为真正的服务运行而非一个shell时(不论是前台服务还是后台daemon)都需要首先创建一个秒盘，这个秒盘会周期性的调用malloc_trim释放glibc缓存的内存块以免造成内存泄露的假象，若服务写本地日志，则该秒盘还将用来周期性的刷日志(日志组件用了自定义的64k缓冲并采取全缓冲策略以避免过多的调用系统调用从而增加吞吐量，因此需要周期性的刷盘以免在某些情况下日志长时间存放在内存中) ②事件对象event在回调函数中执行完回调后将本地所有信号集清空 ③普通服务节点向registry注册时不再附带ip，registry根据本地tcp_server回调中的ip来确定服务节点所在的ip地址

+++ src/seria_impl.cpp
@@ -50,8 +50,8 @@ namespace crx
     void seria::insert(const char *key, const char *data, size_t len)
     {
         auto impl = std::dynamic_pointer_cast<seria_impl>(m_impl);
-        raw_ref rkey(key, strlen(key));
-        raw_ref rval(data, len);
+        raw_ref rkey(key, (uint32_t)strlen(key));
+        raw_ref rval(data, (uint32_t)len);
         std::pair<raw_ref, raw_ref> pair(rkey, rval);
         impl->m_packer.pack(pair);		//将键值对加入序列化串中
